This branch is a prototype branch for exercising the L1 message topology and exploring scoreboard structure. It also allows for timing of transfer operations using messages as the start and the checkered flag.
There are several ancillary helper scripts in this commit to be
used in this branch only.

I expect, with the exception of some component skels, that this
branch will never be merged with ctrl_iip master...but it will be
easier to move new file versions to 25 separate test machines!