This commit message covers the previous commit on this branch; the previous commit message was nil, because, with 4 open and dirty files, the L1test stand system I was working on sent a system message announcing that it was going down for a maintenance window. So, changes were VERY quickly committed and work continued on a different machine.
This section of work finishes up moving around some pieces of the functionality to fit the
L1 sequence diagrams revealed at the September Pathfinder meeting. It also changes the system from being
CCD specific to using Rafts as the atomic fetch unit with CCDs and ancillary argument to a raft that
allows only a portion of the CCDs in a particular raft to be fetched. The general nightly fetch now
has partial readouts (according to configuration desires) and backlog items that only require a handful
of CCDs to be fetched (if a CCD was dropped on transfer to the Archive or Prompt Processing) can be easily
fetched.