This checkin adds a param to the forwarder config description file for whether to use the DAQ to fetch data or not. This is necessary as there only one raft available from DAQ emulator and it is desired that the system run with multiple forwarders - the rest fetching from their local drive.
It also explores a method for converting list structures in incoming yaml messages into cpp vector types.
It adds a bit of intelligence to the Forwarder.cpp callback handler for new_visit, but this needs the yaml
list conversion working. Saving list conversion code so that htut khine can find it easily and chime in.