Before the working Image Sequence was established and how DM fit into the stream of events, It was expected that a Next_Visit event would play an important role. In addition, it was incorrectly thought that TakeImages information would be needed as well.
While the DMCS will need to be aware of Next Visit information for Prompt
Processing data, it is inessential at this point; the Start Integration event has largely
replaced the role of the Next Visit event...so these methods are being removed, which
drops the line count of the file by about 250 lines and makes it easier to understand.

When full telescope testing begins in earnest a year from now or so, the need for a Next
Visit event will be revisited (no pun intended!); it will certainly be much simpler than
the methods removed today.