Correct method signatures to prepare for update to latest (1.0.1) of Pika.
For the most part, changes required to our code are minor, but took a
little while to figure out because of method signature changes:

SelectConnection (and all connections) no longer support the
stop_ioloop_on_close parameter, because it has been removed from Pika.

The reply_text parameter to the on_channel_closed callback is no
longer reported by Pika, and has been removed.

The channel method "basic_cancel" signature changed the order of the
parameters, but since the parameters are named, we can fix this by
using those names on the method call, and it is backwards-compatible
to pika 0.11.2.

The only incompatible change is to the channel method "basic_consume",
which does not use named parameters.  In order to remain
backwards-compatible during the transition to the new libraries, we
check the version number of the library being used, and switch on
that. Since this is a done only on setup, this should have little
impact.  This should, however, be removed at a later date.

This has been tested using Pika version 0.11.2 and version 1.0.1